rename: Move files to providers for clear fsd path
